Change to quality_scale.yaml
