--- homeassistant/components/airgradient/iqs.json
@@ -1,5 +1,0 @@
-{
-  "IQS001": {
-    "implemented": true
-  }
-}
