+++ homeassistant/components/airgradient/quality_scale.yaml
@@ -1,0 +1,2 @@
+select:
+  - IQS001
