+++ script/hassfest/__main__.py
@@ -18,12 +18,12 @@ from . import (
     dhcp,
     docker,
     icons,
-    iqs,
     json,
     manifest,
     metadata,
     mqtt,
     mypy_config,
+    quality_scale,
     requirements,
     services,
     ssdp,
@@ -41,10 +41,10 @@ INTEGRATION_PLUGINS = [
     dependencies,
     dhcp,
     icons,
-    iqs,
     json,
     manifest,
     mqtt,
+    quality_scale,
     requirements,
     services,
     ssdp,
